Added NMS Support for 1.10.X - 1.15.X

+++ MinecraftMediaLibrary/src/main/java/com/github/pulsebeat02/minecraftmedialibrary/nms/PacketHandler.java
@@ -18,7 +18,7 @@ public interface PacketHandler {
 
     void unregisterPlayer(Player player);
 
-    boolean isMapRegistered(int id);
+    boolean isMapRegistered(long id);
 
     void unregisterMap(long id);
