+++ v1_16_R1/src/main/java/com/github/pulsebeat02/minecraftmedialibrary/nms/impl/v1_16_R1/NMSMapPacketIntercepter.java
@@ -219,7 +219,7 @@ public class NMSMapPacketIntercepter implements PacketHandler {
     }
 
     @Override
-    public boolean isMapRegistered(final int id) {
+    public boolean isMapRegistered(final long id) {
         return maps.contains(id);
     }
